Move MemorySnapshot to its own file.
